refactor: api_types/docs/...: Extract API-specify types into new file.
This reduces coupling, and also has the intention of isolating types
involved in use of the API from those used in the implementation.

Entry added to developer-file-overview.md for new file.
