+++ zulipterminal/model.py
@@ -20,9 +20,10 @@ from typing import (
 from urllib.parse import urlparse
 
 import zulip
-from typing_extensions import Literal, TypedDict
+from typing_extensions import Literal
 
 from zulipterminal import unicode_emojis
+from zulipterminal.api_types import EditPropagateMode, Event
 from zulipterminal.config.keys import primary_key_for_command
 from zulipterminal.helper import (
     Message,
@@ -39,42 +40,6 @@ from zulipterminal.helper import (
 )
 from zulipterminal.ui_tools.utils import create_msg_box_list
 
-
-class Event(TypedDict, total=False):  # Each Event will only have a subset
-    type: str
-    # typing:
-    sender: Dict[str, Any]  # 'email', ...
-    # typing & reaction:
-    op: str
-    # reaction:
-    user: Dict[str, Any]  # 'email', 'user_id', 'full_name'
-    reaction_type: str
-    emoji_code: str
-    emoji_name: str
-    # reaction & update_message:
-    message_id: int
-    # update_message:
-    rendered_content: str
-    # update_message_flags:
-    messages: List[int]
-    operation: str  # NOTE: deprecated in Zulip 4.0 / ZFL 32 -> 'op'
-    flag: str
-    all: bool
-    # message:
-    message: Message
-    flags: List[str]
-    subject: str
-    # subscription:
-    property: str
-    user_id: int  # Present when a streams subscribers are updated.
-    user_ids: List[int]  # NOTE: replaces 'user_id' in ZFL 35
-    stream_id: int
-    stream_ids: List[int]  # NOTE: replaces 'stream_id' in ZFL 35 for peer*
-    value: bool
-    message_ids: List[int]  # Present when subject of msg(s) is updated
-
-
-EditPropagateMode = Literal['change_one', 'change_all', 'change_later']
 
 OFFLINE_THRESHOLD_SECS = 140
